andriod build bug fix
