+++ internal/ui/ui.go
@@ -25,7 +25,6 @@ import (
 	"github.com/MehmetMHY/ch/pkg/types"
 	"github.com/ledongthuc/pdf"
 	"github.com/lu4p/cat"
-	"github.com/otiai10/gosseract/v2"
 	"github.com/rwcarlsen/goexif/exif"
 	"github.com/tealeg/xlsx/v3"
 	"golang.org/x/net/html"
@@ -959,55 +958,6 @@ func parseFraction(fraction string) float64 {
 	}
 
 	return numerator / denominator
-}
-
-// extractTextFromImage uses Tesseract OCR to extract text from an image
-func (t *Terminal) extractTextFromImage(filePath string) (string, error) {
-	// Check if tesseract is installed
-	if _, err := exec.LookPath("tesseract"); err != nil {
-		return "", fmt.Errorf("tesseract OCR is not installed. Please install it to enable image-to-text extraction")
-	}
-
-	client := gosseract.NewClient()
-	defer client.Close()
-
-	// Set language (default to English, but could be made configurable)
-	err := client.SetLanguage("eng")
-	if err != nil {
-		// If English fails, try without setting language
-		client = gosseract.NewClient()
-		defer client.Close()
-	}
-
-	// Set image source
-	err = client.SetImage(filePath)
-	if err != nil {
-		return "", fmt.Errorf("failed to set image source: %w", err)
-	}
-
-	// Configure OCR settings for better accuracy
-	client.SetVariable("tessedit_char_whitelist", "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789.,!?@#$%^&*()_+-={}[]|\\:;\"'<>/~` ")
-
-	// Extract text
-	text, err := client.Text()
-	if err != nil {
-		return "", fmt.Errorf("OCR extraction failed: %w", err)
-	}
-
-	// Clean up the extracted text
-	cleanedText := strings.TrimSpace(text)
-
-	// Remove excessive whitespace and normalize line breaks
-	lines := strings.Split(cleanedText, "\n")
-	var cleanedLines []string
-	for _, line := range lines {
-		line = strings.TrimSpace(line)
-		if line != "" {
-			cleanedLines = append(cleanedLines, line)
-		}
-	}
-
-	return strings.Join(cleanedLines, "\n"), nil
 }
 
 // isTextFile checks if content is likely from a text file
